Moving go --> _go and debug --> _debug.
Adding an accumulator method to the HostInfo subclass of Analyzer.

Cleaning up the outQ processing.

Switching from passing in a packer before start to giving the constructor a packerQ
from which it only takes on packer, then it runs and exits.

The abstraction barriers around Host, HostInfo, and CrawlStateManager are all messed up.

Needs refactoring...