The good news is that this version of Server.py and CrawlStateManager.py work and are relatively simple.
The bad news is that the abstraction barriers around hostsQ, HostInfo, Host
and maintaining politeness info are all messed up.

Also, this version creates a new HostInfo object for every single URL that it
sees, which is usually duplicative.

The next_time that is polite to fetch is currently calculate in two different places
and both are in Fetcher, which is probably the wrong place.