This is a partial cleanup of CrawlStateManager, which is still broken and not working.
Fetcher does work, at least to some extent.  There is a test which I will commit next, that
exercises Fetcher using the HostRecord and HostFetchRecord classes and templates defined
in CrawlStateManager.

The big change here is moving all politness out of the Fetcher and into the CrawlStateManager.
This simplifies the Fetcher, so it can just pursue bursts of fetching and sending the data
out to the analyzer chain for processing.

This paves the way for managing all politeness in the crawl state and eventually doing
fancy things like accelerating hosts that have rapidly changing content.